Move iceoryx header to shm_types
