+++ src/core/ddsi/CMakeLists.txt
@@ -178,7 +178,9 @@ if(ENABLE_SECURITY)
 endif()
 if(iceoryx_binding_c_FOUND)
   list(APPEND srcs_ddsi shm_init.c shm_sync.c)
+  list(APPEND srcs_ddsi shm_init.c shm_types.c)
   list(APPEND hdrs_private_ddsi shm_sync.h)
+  list(APPEND hdrs_private_ddsi shm_types.h)
 endif()
 
 prepend(hdrs_private_ddsi "${CMAKE_CURRENT_LIST_DIR}/include/dds/ddsi/" ${hdrs_private_ddsi})
